+++ public/css/app.css
@@ -1,1 +1,1 @@
-@import url("https://fonts.googleapis.com/css?family=Arapey:400,400i|Lato:300,400,700");button,.button{appearance:none;background-color:#f6ede8;border:0;color:#111;cursor:pointer;display:inline-block;font-family:"Arapey",georgia,serif;font-size:16px;-webkit-font-smoothing:antialiased;font-weight:400;line-height:1;padding:0.5rem 1rem;text-align:center;text-decoration:none;transition:background-color 150ms ease;user-select:none;vertical-align:middle;white-space:nowrap}button:hover,.button:hover{background-color:#a5d6d8}button:focus,.button:focus{outline:3px solid #a5d6d8;outline-offset:2px}button:disabled,.button:disabled{cursor:not-allowed;opacity:0.5}button:disabled:hover,.button:disabled:hover{background-color:#f6ede8}fieldset{background-color:transparent;border:0;margin:0;padding:0}legend{font-weight:400;margin-bottom:0.25rem;padding:0}label{display:block;font-weight:400;margin-bottom:0.25rem}input,select,textarea{display:block;font-family:"Arapey",georgia,serif;font-size:1rem}[type=text]{display:block;font-family:"Arapey",georgia,serif;font-size:1rem;appearance:none;background-color:#fff;border:1px solid #e7d0ba;border-radius:5px;box-shadow:inset 0 1px 3px rgba(0,0,0,0.06);box-sizing:border-box;margin-bottom:0.5rem;padding:0.25rem;transition:border-color 150ms ease;width:100%}[type=text]:hover{border-color:#a5d6d8}[type=text]:focus{border-color:#a5d6d8;box-shadow:inset 0 1px 3px rgba(0,0,0,0.06),0 0 5px rgba(165,214,216,0.7);outline:none}[type=text]:disabled{background-color:#f2f2f2;cursor:not-allowed}[type=text]:disabled:hover{border:1px solid #e7d0ba}[type=text]::placeholder{color:#777}textarea{resize:vertical}[type="checkbox"],[type="radio"]{display:inline;margin-right:0.25rem}[type="file"]{margin-bottom:0.5rem;width:100%}select{margin-bottom:0.5rem;width:100%}[type="checkbox"]:focus,[type="radio"]:focus,[type="file"]:focus,select:focus{outline:3px solid #a5d6d8;outline-offset:2px}*,*::before,*::after{box-sizing:border-box}html,body{height:100%}html{background-color:#fff}body{margin:0}ul,ol{list-style-type:none;margin:0;padding:0}dl{margin:0}dt{font-weight:600;margin:0}dd{margin:0}figure{margin:0}img,picture{height:auto;margin:0;max-width:100%}html{color:#111;font-family:"Arapey",georgia,serif;font-size:100%;line-height:1.6}h1,h2,h3,h4,h5,h6{font-family:"Lato",helvetica,sans-serif;font-weight:400;line-height:1.2;margin:0 0 0.5rem}p{margin:0 0 0.5rem}a{border-bottom:2px solid #a5d6d8;color:#111;text-decoration:none;transition:border 150ms ease}a:hover{border-bottom-color:transparent}a:focus{outline:3px solid #a5d6d8;outline-offset:2px}a img{border-bottom:0}hr{border-bottom:1px solid #e7d0ba;border-left:0;border-right:0;border-top:0;margin:1rem 0}.site-header{padding:1rem}.site-header h1{font-size:2rem}.trips-list{display:grid;list-style-type:none;grid-gap:1rem;grid-template-columns:1fr 1fr;margin-left:auto;margin-right:auto;max-width:1000px}.trips-list__link{border-bottom:0;display:block;padding:1rem;transition-property:transform}.trips-list__link:hover{transform:translateY(-0.5rem)}.trips-list__title{display:block;font-size:1.5rem;margin-bottom:0.5rem}p{margin-bottom:0.5rem;margin-top:0.5rem}
+@import url("https://fonts.googleapis.com/css?family=Arapey:400,400i|Lato:300,400,700");button,.button{appearance:none;background-color:#f6ede8;border:0;color:#111;cursor:pointer;display:inline-block;font-family:"Arapey",georgia,serif;font-size:16px;-webkit-font-smoothing:antialiased;font-weight:400;line-height:1;padding:0.5rem 1rem;text-align:center;text-decoration:none;transition:background-color 150ms ease;user-select:none;vertical-align:middle;white-space:nowrap}button:hover,.button:hover{background-color:#a5d6d8}button:focus,.button:focus{outline:3px solid #a5d6d8;outline-offset:2px}button:disabled,.button:disabled{cursor:not-allowed;opacity:0.5}button:disabled:hover,.button:disabled:hover{background-color:#f6ede8}fieldset{background-color:transparent;border:0;margin:0;padding:0}legend{font-weight:400;margin-bottom:0.25rem;padding:0}label{display:block;font-weight:400;margin-bottom:0.25rem}input,select,textarea{display:block;font-family:"Arapey",georgia,serif;font-size:1rem}[type=text]{display:block;font-family:"Arapey",georgia,serif;font-size:1rem;appearance:none;background-color:#fff;border:1px solid #e7d0ba;border-radius:5px;box-shadow:inset 0 1px 3px rgba(0,0,0,0.06);box-sizing:border-box;margin-bottom:0.5rem;padding:0.25rem;transition:border-color 150ms ease;width:100%}[type=text]:hover{border-color:#a5d6d8}[type=text]:focus{border-color:#a5d6d8;box-shadow:inset 0 1px 3px rgba(0,0,0,0.06),0 0 5px rgba(165,214,216,0.7);outline:none}[type=text]:disabled{background-color:#f2f2f2;cursor:not-allowed}[type=text]:disabled:hover{border:1px solid #e7d0ba}[type=text]::placeholder{color:#777}textarea{resize:vertical}[type="checkbox"],[type="radio"]{display:inline;margin-right:0.25rem}[type="file"]{margin-bottom:0.5rem;width:100%}select{margin-bottom:0.5rem;width:100%}[type="checkbox"]:focus,[type="radio"]:focus,[type="file"]:focus,select:focus{outline:3px solid #a5d6d8;outline-offset:2px}*,*::before,*::after{box-sizing:border-box}html,body{height:100%}html{background-color:#fff}body{margin:0}ul,ol{list-style-type:none;margin:0;padding:0}dl{margin:0}dt{font-weight:600;margin:0}dd{margin:0}figure{margin:0}img,picture{height:auto;margin:0;max-width:100%}html{color:#111;font-family:"Arapey",georgia,serif;font-size:100%;line-height:1.6}h1,h2,h3,h4,h5,h6{font-family:"Lato",helvetica,sans-serif;font-weight:400;line-height:1.2;margin:0 0 0.5rem}p{margin:0 0 0.5rem}a{border-bottom:2px solid #a5d6d8;color:#111;text-decoration:none;transition:border 150ms ease}a:hover{border-bottom-color:transparent}a:focus{outline:3px solid #a5d6d8;outline-offset:2px}a img{border-bottom:0}hr{border-bottom:1px solid #e7d0ba;border-left:0;border-right:0;border-top:0;margin:1rem 0}.site-header{padding:1rem}.site-header h1{font-size:2rem}.trips-list{display:grid;list-style-type:none;grid-gap:1rem;grid-template-columns:1fr 1fr;margin-left:auto;margin-right:auto;max-width:1000px}.trips-list__link{border-bottom:0;display:block;padding:1rem;transition-property:transform}.trips-list__link:hover{transform:translateY(-0.5rem)}.trips-list__title{display:block;font-size:1.5rem;margin-bottom:0.5rem}
