+++ public/index.html
@@ -21,7 +21,7 @@
       integrity="sha256-hwg4gsxgFZhOsEEamdOYGBf13FyQuiTwlAQgxVSNgt4="
       crossorigin="anonymous">    
     </script>
-    <script src="/moment.js"></script>
-    <script src="/trips.js"></script>
+    <script src="js/moment.js"></script>
+    <script src="js/trips.js"></script>
   </body>
 </html>
