+++ scss/app.scss
@@ -1,3 +1,2 @@
-@import "variables";
 @import "base/all";
 @import "components/all";
