+++ scss/base/_all.scss
@@ -1,3 +1,5 @@
+@import "variables";
+
 @import "buttons";
 @import "forms";
 @import "layout";
